split API type into single Handler groups
