+++ backend/src/Server/Handlers/DocumentHandlers.hs
@@ -1,26 +1,22 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Server.Handlers.DocumentHandlers (
-      getDocumentHandler
-    , deleteDocumentHandler
-    , getAllExternalUsersDocumentHandler
-    , getExternalUserDocumentHandler
-    , postExternalUserDocumentHandler
-    , deleteExternalUserDocumentHandler
-) where
+module Server.Handlers.DocumentHandlers
+    ( DocumentAPI
+    , documentServer
+    ) where
 
 import Control.Monad.IO.Class
 import Hasql.Connection (Connection)
 import qualified Hasql.Session as Session
 import Servant
 import Servant.Auth.Server
+import Server.Auth (AuthMethod)
 import qualified Server.Auth as Auth
 import Server.HandlerUtil
 import qualified UserManagement.Document as Document
@@ -28,6 +24,49 @@ import qualified UserManagement.Sessions as Sessions
 import qualified UserManagement.User as User
 import VersionControl.Commit
 import Prelude hiding (readFile)
+
+type DocumentAPI =
+    Auth AuthMethod Auth.Token
+        :> "documents"
+        :> Capture "documentID" Document.DocumentID
+        :> Get '[JSON] ExistingCommit
+        :<|> Auth AuthMethod Auth.Token
+            :> "documents"
+            :> Capture "documentID" Document.DocumentID
+            :> Delete '[JSON] NoContent
+        :<|> Auth AuthMethod Auth.Token
+            :> "documents"
+            :> Capture "documentID" Document.DocumentID
+            :> "external"
+            :> Get '[JSON] [(User.UserID, Document.DocPermission)]
+        :<|> Auth AuthMethod Auth.Token
+            :> "documents"
+            :> Capture "documentID" Document.DocumentID
+            :> "external"
+            :> Capture "userID" User.UserID
+            :> Get '[JSON] (Maybe Document.DocPermission)
+        :<|> Auth AuthMethod Auth.Token
+            :> "documents"
+            :> Capture "documentID" Document.DocumentID
+            :> "external"
+            :> Capture "userID" User.UserID
+            :> ReqBody '[JSON] Document.DocPermission
+            :> Post '[JSON] NoContent
+        :<|> Auth AuthMethod Auth.Token
+            :> "documents"
+            :> Capture "documentID" Document.DocumentID
+            :> "external"
+            :> Capture "userID" User.UserID
+            :> Delete '[JSON] NoContent
+
+documentServer :: Server DocumentAPI
+documentServer =
+    getDocumentHandler
+        :<|> deleteDocumentHandler
+        :<|> getAllExternalUsersDocumentHandler
+        :<|> getExternalUserDocumentHandler
+        :<|> postExternalUserDocumentHandler
+        :<|> deleteExternalUserDocumentHandler
 
 getDocumentHandler
     :: AuthResult Auth.Token -> Document.DocumentID -> Handler ExistingCommit
