+++ backend/src/Server/Handlers/RoleHandlers.hs
@@ -1,31 +1,64 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Server.Handlers.RoleHandlers (
-      getRoleHandler
-    , postRoleHandler
-    , deleteRoleHandler
-    , postSuperadminHandler
-    , deleteSuperadminHandler
-) where
+module Server.Handlers.RoleHandlers
+    ( RoleAPI
+    , roleServer
+    ) where
 
 import Control.Monad.IO.Class
 import Hasql.Connection (Connection)
 import qualified Hasql.Session as Session
 import Servant
 import Servant.Auth.Server
+import Server.Auth (AuthMethod)
 import qualified Server.Auth as Auth
 import Server.HandlerUtil
 import qualified UserManagement.Group as Group
 import qualified UserManagement.Sessions as Sessions
 import qualified UserManagement.User as User
 import Prelude hiding (readFile)
+
+type RoleAPI =
+    Auth AuthMethod Auth.Token
+        :> "roles"
+        :> Capture "groupID" Group.GroupID
+        :> Capture "userId" User.UserID
+        :> Get '[JSON] User.Role
+        :<|> Auth AuthMethod Auth.Token
+            :> "roles"
+            :> Capture "groupID" Group.GroupID
+            :> Capture "userId" User.UserID
+            :> ReqBody '[JSON] User.Role
+            :> Put '[JSON] NoContent
+        :<|> Auth AuthMethod Auth.Token
+            :> "roles"
+            :> Capture "groupID" Group.GroupID
+            :> Capture "userId" User.UserID
+            :> Delete '[JSON] NoContent
+        :<|> Auth AuthMethod Auth.Token
+            :> "roles"
+            :> "superadmin"
+            :> Capture "userId" User.UserID
+            :> Post '[JSON] NoContent
+        :<|> Auth AuthMethod Auth.Token
+            :> "roles"
+            :> "superadmin"
+            :> Capture "userId" User.UserID
+            :> Delete '[JSON] NoContent
+
+roleServer :: Server RoleAPI
+roleServer =
+    getRoleHandler
+        :<|> postRoleHandler
+        :<|> deleteRoleHandler
+        :<|> postSuperadminHandler
+        :<|> deleteSuperadminHandler
 
 getRoleHandler
     :: AuthResult Auth.Token -> Group.GroupID -> User.UserID -> Handler User.Role
